+++ slices/equal_test.go
@@ -8,6 +8,20 @@ import (
 	"testing"
 )
 
+func Test_EqualSlice(t *testing.T) {
+
+	{
+		assert.True(t, slices.EqualSlice([]int{}, []int{}))
+		assert.True(t, slices.EqualSlice([]int{1}, []int{1}))
+		assert.False(t, slices.EqualSlice([]int{1}, []int{2}))
+
+		assert.True(t, slices.EqualSlice([]int{1, 2, 3, 4}, []int{1, 2, 3, 4}))
+		assert.False(t, slices.EqualSlice([]int{1, 2, 3, 4}, []int{1, 2, 4, 3}))
+
+		assert.True(t, slices.EqualSlice([]int{1, 2, 3, 4, 5}[:4], []int{1, 2, 3, 4, 6}[:4]))
+	}
+}
+
 func Test_EqualSliceOfInt(t *testing.T) {
 
 	{