+++ CHANGES.md
@@ -3,7 +3,7 @@
 ## 0.10.0-alpha1 - 7th February 2026
 
 * simple refactoring of `ASCIIToLower()` and `ASCIIToUpper()`;
-* changed signature of `CollectSlice()`, including making be a generic function (`CollectSlice[T]()`);
+* changed signature of `CollectSlice()`, including making be a generic function (`CollectSlice[T, U]()`);
 * refactor `CollectSliceOfInt()` in terms of `CollectSliceOfInteger[T]()`;
 * refactored `EqualSliceOfInt()` in terms of `EqualSliceOfInteger[T]()`;
 * refactored `EqualSliceOfUint()` in terms of `EqualSliceOfInteger[T]()`;